refactor utils dir to add create-breakpoints to it
