--- src/utils.js
@@ -1,9 +1,0 @@
-// @flow
-
-import { type Theme } from './theme.types'
-
-export const fromTheme = (key: string) => ({
-  theme,
-}: {
-  theme: Theme,
-}): string => theme[key]
