+++ src/utils/from-theme.js
@@ -1,0 +1,6 @@
+// @flow
+
+import { type Theme } from '../theme.types'
+
+export default (key: string) => ({ theme }: { theme: Theme }): string =>
+  theme[key]
